==> 01_Notebooks/02_Data_Munging/01_Process_V1_Json_to_V1_DF.ipynb <==
### 01 Process V1 Jsons to V1_Event_Fight_Fighter_Fighterinfo

The purpose of this notebook is to convert the V1 Jsons into a dataframe

## Imports

import os
import json
import pandas as pd

## Functions

def Process_v1_jsons_fighter_fights(rootdir):
    list_fighter_fights_dicts = []
    for subdir, dirs, files in os.walk(rootdir):
        for file in files:
            filepath = subdir + os.sep + file
            with open(filepath) as f:
                jsn = json.load(f)

            sub_list_fighter_fights_dict = create_fighter_fights_list_of_dict(jsn)
            list_fighter_fights_dicts.extend(sub_list_fighter_fights_dict)
    return list_fighter_fights_dicts

def create_fighter_fights_list_of_dict(jsn):
    list_of_fighter_fights_dict = []
    list_of_fights = jsn['FMLiveFeed']['Fights']
    for fight in list_of_fights:
        fighter_fights_dict = create_fighter_fights_dictionary(jsn, fight)
        list_of_fighter_fights_dict.append(fighter_fights_dict)
        
    return list_of_fighter_fights_dict

def create_fighter_fights_dictionary(jsn, fight):
    fighter_fights_dict = {}

    # Fight Info
    EventID = try_to_int(jsn['FMLiveFeed']['EventID'])
    FightID = try_to_int(fight['FightID'])
    PossibleRds = try_to_int(fight['PossibleRds'])
    EndingRoundNum = try_to_int(fight['EndingRoundNum'])
    Method = fight['Method']
    

    # Fighter 1 Fight Info
    fighter = fight['Fighters'][0]
    F1_FighterID = try_to_int(fighter['FighterID'])
    F1_Color = fighter['Color']
    F1_FullName = fighter['FullName']
    F1_DOB = fighter['DOB']
    F1_Height = try_to_int(fighter['Height'])
    F1_Weight = fighter['Weight']
    F1_Stance = fighter['Stance']
    F1_Outcome = fighter['Outcome']
    F1_Born_Country = fighter['Born']['Country']
    F1_Born_State = fighter['Born']['State']
    F1_Born_City = fighter['Born']['City']
    F1_Fights_Out_Country = fighter['FightsOutOf']['Country']
    F1_Fights_Out_State = fighter['FightsOutOf']['State']
    F1_Fights_Out_City = fighter['FightsOutOf']['City']
    F1_TotFights = try_to_int(fighter['Record']['TotFights'])
    F1_Win = try_to_int(fighter['Record']['Win'])
    F1_Loss = try_to_int(fighter['Record']['Loss'])
    F1_Draw = try_to_int(fighter['Record']['Draw'])
    F1_NC = try_to_int(fighter['Record']['NC'])

    # Fighter 2 Fight Info
    fighter = fight['Fighters'][1]
    F2_FighterID = try_to_int(fighter['FighterID'])
    F2_Color = fighter['Color']
    F2_FullName = fighter['FullName']
    F2_DOB = fighter['DOB']
    F2_Height = try_to_int(fighter['Height'])
    F2_Weight = fighter['Weight']
    F2_Stance = fighter['Stance']
    F2_Outcome = fighter['Outcome']
    F2_Born_Country = fighter['Born']['Country']
    F2_Born_State = fighter['Born']['State']
    F2_Born_City = fighter['Born']['City']
    F2_Fights_Out_Country = fighter['FightsOutOf']['Country']
    F2_Fights_Out_State = fighter['FightsOutOf']['State']
    F2_Fights_Out_City = fighter['FightsOutOf']['City']
    F2_TotFights = try_to_int(fighter['Record']['TotFights'])
    F2_Win = try_to_int(fighter['Record']['Win'])
    F2_Loss = try_to_int(fighter['Record']['Loss'])
    F2_Draw = try_to_int(fighter['Record']['Draw'])
    F2_NC = try_to_int(fighter['Record']['NC'])
    
    #Add everything to a dictionary 
    fighter_fights_dict['EventID'] = EventID
    fighter_fights_dict['FightID'] = FightID
    fighter_fights_dict['PossibleRds'] = PossibleRds
    fighter_fights_dict['EndingRoundNum'] = EndingRoundNum
    fighter_fights_dict['Method'] = Method
    
    fighter_fights_dict['F1_FighterID'] = F1_FighterID
    fighter_fights_dict['F1_Color'] = F1_Color
    fighter_fights_dict['F1_FullName'] = F1_FullName
    fighter_fights_dict['F1_DOB'] = F1_DOB
    fighter_fights_dict['F1_Height'] = F1_Height
    fighter_fights_dict['F1_Weight'] = F1_Weight
    fighter_fights_dict['F1_Stance'] = F1_Stance
    fighter_fights_dict['F1_Outcome'] = F1_Outcome
    fighter_fights_dict['F1_Born_Country'] = F1_Born_Country
    fighter_fights_dict['F1_Born_State'] = F1_Born_State
    fighter_fights_dict['F1_Born_City'] = F1_Born_City
    fighter_fights_dict['F1_Fights_Out_Country'] = F1_Fights_Out_Country
    fighter_fights_dict['F1_Fights_Out_State'] = F1_Fights_Out_State
    fighter_fights_dict['F1_Fights_Out_City'] = F1_Fights_Out_City
    fighter_fights_dict['F1_TotFights'] = F1_TotFights
    fighter_fights_dict['F1_Win'] = F1_Win
    fighter_fights_dict['F1_Loss'] = F1_Loss
    fighter_fights_dict['F1_Draw'] = F1_Draw
    fighter_fights_dict['F1_Win'] = F1_Win
    fighter_fights_dict['F1_NC'] = F1_NC

    fighter_fights_dict['F2_FighterID'] = F2_FighterID
    fighter_fights_dict['F2_Color'] = F2_Color
    fighter_fights_dict['F2_FullName'] = F2_FullName
    fighter_fights_dict['F2_DOB'] = F2_DOB
    fighter_fights_dict['F2_Height'] = F2_Height
    fighter_fights_dict['F2_Weight'] = F2_Weight
    fighter_fights_dict['F2_Stance'] = F2_Stance
    fighter_fights_dict['F2_Outcome'] = F2_Outcome
    fighter_fights_dict['F2_Born_Country'] = F2_Born_Country
    fighter_fights_dict['F2_Born_State'] = F2_Born_State
    fighter_fights_dict['F2_Born_City'] = F2_Born_City
    fighter_fights_dict['F2_Fights_Out_Country'] = F2_Fights_Out_Country
    fighter_fights_dict['F2_Fights_Out_State'] = F2_Fights_Out_State
    fighter_fights_dict['F2_Fights_Out_City'] = F2_Fights_Out_City
    fighter_fights_dict['F2_TotFights'] = F2_TotFights
    fighter_fights_dict['F2_Win'] = F2_Win
    fighter_fights_dict['F2_Loss'] = F2_Loss
    fighter_fights_dict['F2_Draw'] = F2_Draw
    fighter_fights_dict['F2_Win'] = F2_Win
    fighter_fights_dict['F2_NC'] = F2_NC
    
    return fighter_fights_dict

def try_to_int(field):
    if field is not None:
        return int(field)

## Run the Functions

rootdir = '../../02_Data/01_Raw_Scraped_Data/V1_Jsons/'
list_figher_fights_dict = Process_v1_jsons_fighter_fights(rootdir)
df = pd.DataFrame(list_figher_fights_dict)

df.head(2)

## Setup Column Order: [Other Columns] [F1 Columns] [F2 Columns]

F1_Columns = [col for col in df.columns if 'F1' in col]
F2_Columns = [col for col in df.columns if 'F2' in col]
Other_Columns = [col for col in df.columns if not 'F2' in col and not 'F1' in col]
Other_Columns = ['EventID', 'FightID', 'Method', 'PossibleRds','EndingRoundNum']

F1_Columns.pop(F1_Columns.index('F1_Outcome'))
F1_Columns.pop(F1_Columns.index('F1_FullName'))
F2_Columns.pop(F2_Columns.index('F2_Outcome'))
F2_Columns.pop(F2_Columns.index('F2_FullName'))

Ordered_Columns = Other_Columns + ['F1_FullName','F2_FullName','F1_Outcome','F2_Outcome'] + F1_Columns + F2_Columns

df = df[Ordered_Columns]

## Duplicate and flip the data **

# Setup Columns Again
F1_Columns = [col for col in df.columns if 'F1' in col]
F2_Columns = [col for col in df.columns if 'F2' in col]
Other_Columns = [col for col in df.columns if not 'F2' in col and not 'F1' in col]
Other_Columns = ['EventID', 'FightID', 'Method', 'PossibleRds','EndingRoundNum']

Ordered_Columns = Other_Columns + F1_Columns + F2_Columns
Flipped_Columns = Other_Columns + F2_Columns + F1_Columns

# Put Columns in Order
df = df[Ordered_Columns]

# Create Flipped df
flipped_df = df[Flipped_Columns]
flipped_df.columns = Ordered_Columns

# Concatenate df and flipped_df
df_with_flipped = pd.concat([df, flipped_df])

# Put the columns back in the other order
F1_Columns = [col for col in df_with_flipped.columns if 'F1' in col]
F2_Columns = [col for col in df_with_flipped.columns if 'F2' in col]
Other_Columns = [col for col in df_with_flipped.columns if not 'F2' in col and not 'F1' in col]
Other_Columns = ['EventID', 'FightID', 'Method', 'PossibleRds','EndingRoundNum']

F1_Columns.pop(F1_Columns.index('F1_Outcome'))
F1_Columns.pop(F1_Columns.index('F1_FullName'))
F2_Columns.pop(F2_Columns.index('F2_Outcome'))
F2_Columns.pop(F2_Columns.index('F2_FullName'))

Ordered_Columns = Other_Columns + ['F1_FullName','F2_FullName','F1_Outcome','F2_Outcome'] + F1_Columns + F2_Columns

df_with_flipped = df_with_flipped[Ordered_Columns]

## Export To CSV

df.to_csv('../../02_Data/02_Processed_Data/V1_Fight_Fighters_FighterInfo.csv')
df_with_flipped.to_csv('../../02_Data/02_Processed_Data/V1_Fight_Fighters_FighterInfo_w_flipped.csv')


!ls ../../02_Data/02_Processed_Data/
