==> 01_Notebooks/02_Data_Munging/03_Transform_V2_Fight_Details_Add_Percent_and_Diff.ipynb <==
# 03 Transform V2 Fight Details to add Percentages and Differentials
- Calculate historical average of fight stats for each fight
- Everything is from the perspective of the F1 fighter
- Stats for the F1 fighter are basically an "offensive" rating for the F1 fighter
- Stats for the F2 fighter are basically an inverse "defensive" rating for the F1 fighter
- We also want to look at the difference in the stats for each fight, and then historical average difference

## Imports

import pandas as pd
import re

## Pull in V2_Fight_Details CSV

df = pd.read_csv('../../02_Data/02_Processed_Data/V2_Fight_Details.csv',index_col=0)

#### Notes on this initial set of data:
- No nulls currently since I dropped them all already
- No time in position details since they had nulls
- Only 1 side.  Not flipped and appended yet

## Need to add and flip before the transformations

# Setup Columns Again
F1_Columns = [col for col in df.columns if 'f1' in col.lower()]
F2_Columns = [col for col in df.columns if 'f2' in col.lower()]
Other_Columns = [col for col in df.columns if not 'f2' in col.lower() and not 'f1' in col.lower()]

Ordered_Columns = Other_Columns + F1_Columns + F2_Columns
Flipped_Columns = Other_Columns + F2_Columns + F1_Columns

# Put Columns in Order
df = df[Ordered_Columns]

# Create Flipped df
flipped_df = df[Flipped_Columns]
flipped_df.columns = Ordered_Columns

# Concatenate df and flipped_df
df = pd.concat([df, flipped_df])

# Reset the index
df = df.reset_index().drop(columns=['index'])

## Setup Column References for Easy Access  *Do not delete*

# These columns are used in the transformations
F1_Columns = [col for col in df.columns if 'f1' in col.lower()]
F2_Columns = [col for col in df.columns if 'f2' in col.lower()]
Other_Columns = [col for col in df.columns if not 'f2' in col.lower() and not 'f1' in col.lower()]

F1_Strikes = [col for col in F1_Columns if 'strikes' in col.lower()] + ['F1_Knock_Down_Landed']
F1_Grappling = [col for col in F1_Columns if 'grappling' in col.lower()]
F1_TIP = [col for col in F1_Columns if 'tip' in col.lower()]
F1_Identification = ['F1_FighterID','F1_Name'] # This is what its supposed to be incase I add more columns
F1_Identification = list(set(F1_Columns) - set(F1_Strikes) - set(F1_Grappling) - set(F1_TIP))

F2_Strikes = [col for col in F2_Columns if 'strikes' in col.lower()] + ['F2_Knock_Down_Landed']
F2_Grappling = [col for col in F2_Columns if 'grappling' in col.lower()]
F2_TIP = [col for col in F2_Columns if 'tip' in col.lower()]
F2_Identification = ['F2_FighterID','F2_Name']
F2_Identification = list(set(F2_Columns) - set(F2_Strikes) - set(F2_Grappling) - set(F2_TIP))

## Perform Transformation:  Calculate success percent for every pair of words that have "landed" and "attempted"

def Perform_calc_success_percent(df):
    Both_Landed_Attempts, Attempts_Only, Landed_Only = get_attempts_landed_columns(df)
    num_columns_before = df.shape[1]
    columns_before = df.columns
    print(f'num columns before: {num_columns_before}')
    for col in Both_Landed_Attempts:
        df = calc_success_percent(df, col)
    num_columns_after = df.shape[1]
    columns_after = df.columns
    print(f"# Columns Added: {num_columns_after-num_columns_before}")
    print(f'num columns after: {num_columns_after}')
    new_columns = list(set(columns_after) - set(columns_before))
    return df, new_columns

def calc_success_percent(df, col_root):
    attempts = col_root + '_attempts'
    landed = col_root + '_landed'
    success_percent = col_root + '_percent'
    
    df[success_percent] = df.apply(lambda x: x[landed]/x[attempts] if x[attempts] != 0 else 0, axis=1)
    
    return df

#calc_success_percent(df, 'F1_Body_Significant_Strikes')

def get_attempts_landed_columns(df):
    list_cols_attempts = []
    list_cols_landed = []
    for column in df.columns: 
        if '_attempts' in column.lower():
            list_cols_attempts.append(column)
        if '_landed' in column.lower():
            list_cols_landed.append(column)
    list_cols_attempts = [re.sub('_attempts','',col) for col in list_cols_attempts] # Remove "_Attempts"
    list_cols_landed = [re.sub('_landed','',col) for col in list_cols_landed]
    Attempts_Only = set(list_cols_attempts) - set(list_cols_landed)
    Landed_Only =  set(list_cols_landed) - set(list_cols_attempts)
    Both_Landed_Attempts = set(list_cols_landed) & set(list_cols_attempts)
    return Both_Landed_Attempts, Attempts_Only, Landed_Only

### Run the Function

df, Landed_Percent = Perform_calc_success_percent(df)

## Perform Transformation:  Calculate In Fight Differentials

f1_calc_columns = list(set(F1_Columns) - set(F1_Identification) - set(F1_TIP))

# I also need to get the percent columns
# At this point the df already has the percent columns... so just grab them from the df directly
percent_cols = [col[3:] for col in df.columns if 'percent' in col and 'f1' in col]

def Perform_calc_fight_stat_differential(df):
    percent_cols = [col[3:] for col in df.columns if 'percent' in col and 'f1' in col]
    f1_calc_columns = list(set(F1_Columns) - set(F1_Identification) - set(F1_TIP))
    for_calc_columns = [col[3:] for col in f1_calc_columns] # Remove the F1_ part
    for_calc_columns.extend(percent_cols)
    num_columns_before = df.shape[1]
    columns_before = df.columns
    print(f'num columns before: {num_columns_before}')
    for col in for_calc_columns:
        df = calc_fight_stat_differential(df, col)
    num_columns_after = df.shape[1]
    columns_after = df.columns
    print(f"# Columns Added: {num_columns_after-num_columns_before}")
    print(f'num columns after: {num_columns_after}')
    new_columns = list(set(columns_after) - set(columns_before))
    return df, new_columns

def calc_fight_stat_differential(df, col_root):
    f1 = 'f1_' + col_root
    f2 = 'f2_' + col_root
    diff = col_root + '_diff'
    df[diff] = df.apply(lambda x: (x[f1] - x[f2])/(x[f1] + x[f2]) if (x[f1] + x[f2]) != 0 else 0, axis=1)
    return df

# calc_fight_stat_differential(df, col_root)

### Run the Function

df, stat_diff = Perform_calc_fight_stat_differential(df)

## This dataset is currently missing the date.  Join it from the Events_df

path = '../../02_Data/01_Raw_Scraped_Data/Events/events_df.csv'

events_df = pd.read_csv(path, index_col=0)
events_df = events_df[['EventId','Date']]
events_df = events_df.rename(index=str, columns={"EventId": "eventid", 'Date':'date'})

df = df.merge(events_df, on='eventid')

df.shape

df.head()

## Processing for V2 Df is done.  Export it

df.to_csv('../../02_Data/02_Processed_Data/V2_Processed.csv')

!ls ../../02_Data/02_Processed_Data/

# Training Stub Data
train_stub = ['eventid','fightid','f1_fullname','f2_fullname','f1_fighterid','f2_fighterid','f1_outcome']
train = V1_df[train_stub].reset_index().drop(columns='index')


df_ems_f1 = df_ems.copy()

# First append 'F1_' for all fighter 1 data
df_ems_f1.columns = ['eventid','fightid','f1_fighterid','date'] + \
                    ['f1_' + col for col in df_ems_f1.columns if col not in \
                    ['eventid','fightid','f1_fighterid','date']]

# Merge F1 Expanding Means
train = train.merge(df_ems_f1, left_on=['eventid','fightid','f1_fighterid'],
                    right_on=['eventid','fightid','f1_fighterid'])

#Setup 
df_ems_f2 = df_ems.drop(columns=['date']).copy()
df_ems_f2.columns = ['eventid','fightid','f2_fighterid'] + \
                    ['f2_' + col for col in df_ems_f2.columns if col not in \
                    ['eventid','fightid','f1_fighterid']]

# Merge em for fighter 2
train = train.merge(df_ems_f2, left_on=['eventid','fightid','f2_fighterid'], 
                    right_on=['eventid','fightid','f2_fighterid'])

# drop columns w/o outcome
train = train.dropna(axis=0, how='any')

# Check empty columns:
check_nulls(train)

# Need to clean up data a bit more before I can model
- Label from win/loss to 1/0

train.f1_outcome.value_counts()

train['outcome'] = train.f1_outcome.map(lambda x: 1 if x=='Win' else 0)

# Export Train dataset for some EDA

train.to_csv('../../02_Data/02_Processed_Data/train.csv')
