==> 01_Notebooks/02_Data_Munging/04_Transform_V2_Processed_to_Historical_Averages.ipynb <==
# 03 Transform V2 Processed into Historical Averages

## Imports

import pandas as pd
import re

## Pull in V2 Processed DataFrame

df = pd.read_csv('../../02_Data/02_Processed_Data/V2_Processed.csv', index_col=0)

df.shape

def calculate_expanding_mean(df):
    # Find the unique ID for every fighter
    unique_ids = df.f1_fighterid.unique()
    
    cols_to_calc_em = df.select_dtypes(exclude='object').columns
    cols_to_calc_em = [col for col in cols_to_calc_em if col not in ['eventid','fightid','currentrnd','maxrnds',
                                                                     'f1_fighterid','f2_fighterid']]
    
    for count, unique_id in enumerate(unique_ids):
        new_df = df[(df.f1_fighterid == unique_id)]
        if count%50==0:
            print(f"{count} in {len(unique_ids)}")
        
        
        # Calculate Expanding Mean
        df_stub = new_df.sort_values('date')[['eventid','fightid', 'f1_fighterid', 'date']].copy()
        df_em_part = new_df.sort_values('date').expanding().mean().shift()[cols_to_calc_em]
        df_em_part.columns = [col +'_avg' for col in df_em_part.columns]
        df_em = pd.concat([df_stub,df_em_part], axis=1).copy()        
        df_em = df_em.iloc[1:, :]
        
        try:
            df_ems = pd.concat([df_ems, df_em], axis=0)

        except NameError:
            df_ems = df_em.copy() # If the dataframe doesn't exist yet, then copy it from df_em
            
        
        
    return df_ems

historical_avg = calculate_expanding_mean(df)

historical_avg.shape

## Export this and just first join back

historical_avg.to_csv('../../02_Data/02_Processed_Data/historical_avgs.csv')

## Look at this file and see if we can change the names here

historical_avg = pd.read_csv('../../02_Data/02_Processed_Data/historical_avgs.csv', index_col=0)

for col in historical_avg.columns:
    print(col)
