==> 01_Notebooks/02_Data_Munging/05_Cleanse_Fighter_Page_Info.ipynb <==
# 04 Cleanse Fighter Page Info

## Cleansing TODOs:
- Convert reach string -> inches
- Convert weight string -> lbs
- Convert height string -> inches
- Convert fighter slug -> all lower case, no spaces or dashs

## Main Goal:
- Be able to join this data to the training stub

## Imports

import pandas as pd
import re

## Bring in raw data frame

Fighter_Info2181 = pd.read_csv('../../02_Data/01_Raw_Scraped_Data/UFC_Fighter_Page_Info2181.csv', index_col=0)

# Clean up column names
Fighter_Info2181.columns = ['age', 'college', 'degree', 'fights_out', 'from', 'height',
       'leg_reach', 'nickname', 'past', 'reach', 'weight', 'fighter_slug']

# Select relevant columns
df = Fighter_Info2181[['fighter_slug', 'reach', 'weight', 'height', 'leg_reach']].copy()

## Column cleansing

# Convert reach string -> inches
# There are nans in data
df.loc[df.reach.notnull(),'reach'] = df[df.reach.notnull()].reach.map(lambda x: x[:-1])

# Convert leg_reach string -> inches
# There are nans in data
df.loc[df.leg_reach.notnull(),'leg_reach'] = df[df.leg_reach.notnull()].leg_reach.map(lambda x: x[:-1])

# Convert weight to lbs
# There are nans in data
df.loc[df.weight.notnull(),'weight'] = df[df.weight.notnull()].weight.map(lambda x: x[:3])

# convert slug to lower case and no space or dash
df.fighter_slug = df.fighter_slug.map(lambda slug: re.sub('-','',slug).lower())

# Convert height to inches
def convert_height(height_string):
    feet, inches, _ = re.findall("\d+", height_string)
    feet = int(feet)
    inches = int(inches)
    total_inches = feet*12 + inches
    return total_inches

df.loc[df.height.notnull(),'height'] = df[df.height.notnull()].height.map(convert_height)

# Drop leg_reach since too many nans
df = df.drop(columns='leg_reach')

df.head(2)

## Bring in Fighter ID into this dataframe

stub = pd.read_csv('../../02_Data/02_Processed_Data/train_stub.csv', index_col=0)


fightidmap = stub[['f1_fighterid', 'f1_slug', 'eventid']]
fightidmap.columns = ['fighterid', 'fighter_slug', 'eventid']
fightidmap = fightidmap.groupby(['fighterid', 'fighter_slug']).sum().reset_index().drop(columns='eventid')

df = df.merge(fightidmap, how='left', on='fighter_slug')

df.head()

## Export Cleansed Fighter Page Info

df.to_csv('../../02_Data/02_Processed_Data/fighter_page_info.csv')
