==> 01_Notebooks/02_Data_Munging/06_Cleanse_V1_for_merge.ipynb <==
# 05 Cleanse V1 into Fighter Static Info

## Imports

import pandas as pd
import numpy as np
from scipy import stats

## Import V1_DF_w_flipped and convert columns to lowercase

V1_df = pd.read_csv('../../02_Data/02_Processed_Data/V1_DF_w_flipped.csv', index_col=0)
V1_df = V1_df.reset_index().drop(columns='index')
V1_df.columns = [col.lower() for col in V1_df.columns]

# Build out the list of columns I want to pull from V1_df
stats_columns = ['eventid','f1_fighterid', 'f1_height','f1_weight','f1_stance']

# Create a dataframe fighter_stats
fighter_static_stats = V1_df[stats_columns].copy()
# Change Stance to boolean
fighter_static_stats.f1_stance = fighter_static_stats.f1_stance.map(lambda x: 0 if x == 'Orthodox' else 1)
fighter_static_stats.columns = ['eventid']+[col[3:] for col in fighter_static_stats.columns if 'eventid' != col]

## Use fighter page info to join here

fighter_page_info = pd.read_csv('../../02_Data/02_Processed_Data/fighter_page_info.csv', index_col=0)

fighter_page_info.head(2)

There are 220 weights with 0 value in the static fighter stats.  Use data from fighter page info to fill it in

fighter_static_stats[fighter_static_stats.weight == 0].shape

## Fill in 0 weights using fighter page info

def get_fighter_page_info(fighter_id):
    try:
        value = fighter_page_info.loc[fighter_page_info.fighterid==fighter_id,'weight'].iloc[0]
    except:
        value = 0
    
    return value

fighter_static_stats.loc[fighter_static_stats.weight == 0,'weight'] = \
fighter_static_stats[fighter_static_stats.weight == 0]['fighterid'].map(get_fighter_page_info)

After filling, 1 fighter is still 0 weight.  He has a height of 70 inches.  Use the mode of fighters that are 70 inches tall to fill in the value

fighter_static_stats[fighter_static_stats.weight == 0]

fighter_static_stats[fighter_static_stats.weight == 0] = \
fighter_static_stats[fighter_static_stats.height==70].weight.value_counts().keys()[0]

No More 0 weights

fighter_static_stats[fighter_static_stats.weight == 0].shape

## Finally, merge reach from fighter page info onto static fighter stats
- We won't need to merge fighter page info anymore

fighter_page_info.head()

for_merge = fighter_page_info[['fighterid','reach']].dropna()

fighter_static_stats = fighter_static_stats.merge(for_merge, how='left', on='fighterid')

## 107 reach values are null.  Impute the reach using the height

fighter_static_stats.isnull().sum()

fighter_static_stats.loc[fighter_static_stats.reach.isnull(), 'reach'] = \
fighter_static_stats[fighter_static_stats.reach.isnull()].height

## Calculate physical property ratios
- reach to height ratio
- reach to weight ratio
- height to weight ratio

fighter_static_stats['reach_height_ratio'] = fighter_static_stats.apply(lambda x: x['reach']/x['height'], axis=1)
fighter_static_stats['reach_weight_ratio'] = fighter_static_stats.apply(lambda x: x['reach']/x['weight'], axis=1)
fighter_static_stats['height_weight_ratio'] = fighter_static_stats.apply(lambda x: x['height']/x['weight'], axis=1)

## Export fighter static stats - to be merged using eventid and fighterid

fighter_static_stats.to_csv('../../02_Data/02_Processed_Data/fighter_static_stats.csv')
