==> 01_Notebooks/02_Data_Munging/09_Calculate_F1_vs_F2_Differentials_Addl_Processing.ipynb <==
# 09 Calculate F1 vs. F2 Differentials and Addl Processing
- Calculate the F1 and F2 differentials
- Add other feature engineering here
- Train test split here

## Imports

import pandas as pd
import pickle

df = pd.read_csv('../../02_Data/02_Processed_Data/train.csv', index_col=0)

f1_columns = [col for col in df.columns[9:] if col[:2]=='f1']
f2_columns = [col for col in df.columns[9:] if col[:2]=='f2']

def Perform_calc_fight_stat_differential(df):
    f1_columns = [col for col in df.columns[9:] if col[:2]=='f1']
    for_calc_columns = [col[3:] for col in f1_columns] # Remove the F1_ part
    num_columns_before = df.shape[1]
    columns_before = df.columns
    print(f'num columns before: {num_columns_before}')
    for key, col in enumerate(for_calc_columns):
        df = calc_fight_stat_differential(df, col)
        if key%100==0:
            print(key)
    num_columns_after = df.shape[1]
    columns_after = df.columns
    print(f"# Columns Added: {num_columns_after-num_columns_before}")
    print(f'num columns after: {num_columns_after}')
    new_columns = list(set(columns_after) - set(columns_before))
    return df, new_columns

def calc_fight_stat_differential(df, col_root):
    f1 = 'f1_' + col_root
    f2 = 'f2_' + col_root
    diff = col_root + '_diff'
    df[diff] = df.apply(lambda x: (x[f1] - x[f2])/(x[f1] + x[f2]) if (x[f1] + x[f2]) != 0 else 0, axis=1)
    return df

# calc_fight_stat_differential(df, col_root)

df, new_diff = Perform_calc_fight_stat_differential(df)

df.head(2)

## Convert outcomes to a boolean

df['outcome'] = df.f1_outcome.map(lambda x: 1 if x=='Win' else 0)

## Add reach and height advantage

df['f1_reach_adv'] = df.apply(lambda x: x['f1_reach'] - x['f2_reach'], axis=1)
df['f1_height_adv'] = df.apply(lambda x: x['f1_height'] - x['f2_height'], axis=1)

## Train test split

unique_ids = list(set(df.f1_fighterid))
test_full = df[0:0]
train_full = df[0:0]


for fighter_id in unique_ids:
    sub_df = df[df.f1_fighterid == fighter_id].sort_values('date', ascending=False)
#     print(sub_df.shape[0])
#     print(fighter_id)
    if sub_df.shape[0] > 6:
        test_records = sub_df[:2]
        train_records = sub_df[2:]
        test_full = pd.concat([test_full, test_records], axis=0)
        train_full = pd.concat([train_full, train_records], axis=0)
    elif sub_df.shape[0] > 3:
        test_records = sub_df[:1]
        train_records = sub_df[1:]
        test_full = pd.concat([test_full, test_records], axis=0)
        train_full = pd.concat([train_full, train_records], axis=0)
    else:
        train_records = sub_df
        train_full = pd.concat([train_full, train_records], axis=0)

# Split out X and Y

label = 'outcome'
features = [col for col in train_full.columns if col != label]
y_train = train_full[label]
X_train = train_full[features]

# Drop non-number features
X_train = X_train.select_dtypes(exclude='object')
# Drop non-feature columns
X_train = X_train.drop(columns=['eventid','fightid','f1_fighterid','f2_fighterid'])

label = 'outcome'
features = [col for col in test_full.columns if col != label]
y_test = test_full[label]
X_test = test_full[features]

# Drop non-number features
X_test = X_test.select_dtypes(exclude='object')
# Drop non-feature columns
X_test = X_test.drop(columns=['eventid','fightid','f1_fighterid','f2_fighterid'])

X_train.shape, X_test.shape

## Export the Training and Test Data

with open('../../02_Data/02_Processed_Data/X_train.pkl', 'wb') as f:
    pickle.dump(X_train, f)

with open('../../02_Data/02_Processed_Data/y_train.pkl', 'wb') as f:
    pickle.dump(y_train, f)

with open('../../02_Data/02_Processed_Data/X_test.pkl', 'wb') as f:
    pickle.dump(X_test, f)

with open('../../02_Data/02_Processed_Data/y_test.pkl', 'wb') as f:
    pickle.dump(y_test, f)

## Export Train_2

df.to_csv('../../02_Data/02_Processed_Data/train_2.csv')
